requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-google-genai
langchain-cohere
google-generativeai
pymupdf
python-dotenv
docx2txt
psycopg2
fastapi
sqlalchemy
uvicorn
python-jose[cryptography]
passlib[bcrypt]
cohere
pydantic
aiohttp
